messenger_app/requirements.txt
backcall==0.1.0
coverage==5.0.3
decorator==4.4.2
Django==2.2.5
django-cors-headers==3.2.1
django-selenium==0.9.8
djangorestframework==3.11.0
factory-boy==2.12.0
Faker==4.0.2
flamegraph==0.1
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
mock==4.0.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pkg-resources==0.0.0
prompt-toolkit==3.0.3
psycopg2==2.8.4
ptyprocess==0.6.0
Pygments==2.5.2
python-dateutil==2.8.1
python-memcached==1.59
pytz==2019.3
selenium==3.141.0
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.3
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
